Refactor test files
Create aux.jl and move some helper functions there.
Also create a module C to hold constants inside aux.jl.
Create test-* files and move the test sets to those files.
Create more helper functions to improve readibility of the tests, such
as _with_tmp_dir to run mktempdir and cd at the same time and
_basic_new_pkg to create a package with Pkg and run the git setup.

Closes #340
